+++ packages/leafygreen-provider/src/LeafyGreenContext.tsx
@@ -1,26 +1,34 @@
-import React from 'react';
+import React, { PropsWithChildren } from 'react';
 import PropTypes from 'prop-types';
 import UsingKeyboardProvider from './UsingKeyboardContext';
 import TypographyProvider, {
   TypographyProviderProps,
 } from './TypographyContext';
 import PortalContextProvider, { PortalContextValues } from './PortalContext';
+import DarkModeProvider from './DarkModeContext';
+import { DarkModeProps } from '@leafygreen-ui/lib';
 
 type LeafyGreenProviderProps = {
-  children: React.ReactNode;
+  /**
+   * Define a container HTMLElement for components that utilize the `Portal` component
+   */
   popoverPortalContainer?: PortalContextValues['popover'];
-} & TypographyProviderProps;
+} & TypographyProviderProps &
+  DarkModeProps;
 
 function LeafyGreenProvider({
   children,
   baseFontSize,
   popoverPortalContainer,
-}: LeafyGreenProviderProps) {
+  darkMode,
+}: PropsWithChildren<LeafyGreenProviderProps>) {
   return (
     <UsingKeyboardProvider>
       <PortalContextProvider popover={popoverPortalContainer}>
         <TypographyProvider baseFontSize={baseFontSize}>
-          {children}
+          <DarkModeProvider globalDarkMode={darkMode}>
+            {children}
+          </DarkModeProvider>
         </TypographyProvider>
       </PortalContextProvider>
     </UsingKeyboardProvider>
@@ -29,6 +37,15 @@ function LeafyGreenProvider({
 
 LeafyGreenProvider.displayName = 'LeafyGreenProvider';
 
-LeafyGreenProvider.propTypes = { children: PropTypes.node };
+LeafyGreenProvider.propTypes = {
+  popoverPortalContainer: PropTypes.shape({
+    popover: PropTypes.shape({
+      portalContainer: PropTypes.instanceOf(HTMLElement),
+      scrollContainer: PropTypes.instanceOf(HTMLElement),
+    }),
+  }),
+  baseFontSize: PropTypes.oneOf([14, 16]),
+  darkMode: PropTypes.bool,
+};
 
 export default LeafyGreenProvider;