+++ packages/select/src/MenuButton.tsx
@@ -1,23 +1,23 @@
 import React, { useCallback, useContext } from 'react';
-import Button, { Size, Variant } from '@leafygreen-ui/button';
+import Button, { Size as ButtonSize, Variant } from '@leafygreen-ui/button';
 import { css, cx } from '@leafygreen-ui/emotion';
 import CaretDownIcon from '@leafygreen-ui/icon/dist/CaretDown';
 import { breakpoints, spacing } from '@leafygreen-ui/tokens';
 import { palette, uiColors } from '@leafygreen-ui/palette';
 import WarningIcon from '@leafygreen-ui/icon/dist/Warning';
 import { HTMLElementProps } from '@leafygreen-ui/lib';
-import { colorSets, mobileSizeSet, Mode, sizeSets } from './styleSets';
+import { Mode, State } from './types';
+import { colorSets, mobileSizeSet, sizeSets } from './styleSets';
 import SelectContext from './SelectContext';
 import { useForwardedRef } from './utils';
-import { State } from '.';
 
 const menuButtonStyleOverrides = css`
-  text-transform: unset;
-  font-weight: 400;
   // Override button defaults
   > *:last-child {
     grid-template-columns: 1fr 16px;
     padding: 0 12px;
+    justify-content: flex-start;
+
     > svg {
       justify-self: right;
       width: 16px;
@@ -71,6 +71,7 @@ const menuButtonTextWrapperStyle = css`
   align-items: center;
   flex-grow: 1;
   gap: ${spacing[1]}px;
+  overflow: hidden;
 `;
 
 const menuButtonTextStyle = css`
@@ -172,15 +173,15 @@ const MenuButton = React.forwardRef<HTMLElement, Props>(function MenuButton(
   const buttonClassName = __INTERNAL__menuButtonSlot__
     ? ''
     : cx(
-        menuButtonStyleOverrides, // TODO: Refresh - remove overrides
-        menuButtonModeOverrides[mode], // TODO: Refresh - remove overrides
+        menuButtonStyleOverrides,
+        menuButtonModeOverrides[mode],
         {
           [menuButtonDeselectedStyles[mode]]: deselected,
           [menuButtonDisabledStyles[mode]]: disabled,
           [menuButtonErrorStyle[mode]]: state === State.Error && !!errorMessage,
           [css`
             letter-spacing: initial;
-          `]: size === Size.XSmall,
+          `]: size === ButtonSize.XSmall,
         },
         css`
           width: 100%;
@@ -191,6 +192,8 @@ const MenuButton = React.forwardRef<HTMLElement, Props>(function MenuButton(
         `,
       );
 
+  const testId =
+    (rest as any)['data-testid'] ?? 'leafygreen-ui-select-menubutton';
   return (
     <Component
       {...rest}
@@ -203,7 +206,7 @@ const MenuButton = React.forwardRef<HTMLElement, Props>(function MenuButton(
       darkMode={mode === Mode.Dark}
       rightGlyph={<CaretDownIcon />}
       size={size}
-      data-testid="leafygreen-ui-select-menubutton"
+      data-testid={testId}
       className={buttonClassName}
     >
       <div className={menuButtonTextWrapperStyle}>