+++ packages/select/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @leafygreen-ui/select
 
+## 6.1.0
+
+### Minor Changes
+
+- 828c895d: Passes through all valid `button`, `li` and `div` attriutes (including `data-*` attributes) to `Select`, `Option` and `OptionGroup` respectively
+
+### Patch Changes
+
+- b22cc41f: Fixes a bug where long values, when selected, would overflow outside the Select trigger.
+- Updated dependencies [13a4adcc]
+  - @leafygreen-ui/typography@13.0.1
+
+## 6.0.0
+
+### Patch Changes
+
+- Updated dependencies [85d46871]
+- Updated dependencies [99e20bb9]
+  - @leafygreen-ui/lib@9.4.0
+  - @leafygreen-ui/leafygreen-provider@2.3.0
+  - @leafygreen-ui/button@16.0.0
+  - @leafygreen-ui/interaction-ring@4.0.0
+  - @leafygreen-ui/popover@9.0.0
+  - @leafygreen-ui/typography@13.0.0
+
 ## 5.0.3
 
 ### Patch Changes